db/requirements.txt
weaviate-client==4.10.4
pandas==1.5.3
